+++ src/PhpPresentation/DocumentLayout.php
@@ -74,16 +74,14 @@ class DocumentLayout
     private $layout;
 
     /**
-     * Layout x dimension
-     *
-     * @var integer
+     * Layout X dimension
+     * @var float
      */
     private $dimensionX;
 
     /**
-     * Layout y dimension
-     *
-     * @var integer
+     * Layout Y dimension
+     * @var float
      */
     private $dimensionY;
 
@@ -150,7 +148,7 @@ class DocumentLayout
     /**
      * Get Document Layout cx
      *
-     * @param float $unit
+     * @param string $unit
      * @return integer
      */
     public function getCX($unit = self::UNIT_EMU)
@@ -161,7 +159,7 @@ class DocumentLayout
     /**
      * Get Document Layout cy
      *
-     * @param float $unit
+     * @param string $unit
      * @return integer
      */
     public function getCY($unit = self::UNIT_EMU)
@@ -173,8 +171,8 @@ class DocumentLayout
      * Get Document Layout cx
      *
      * @param float $value
-     * @param float $unit
-     * @return integer
+     * @param string $unit
+     * @return DocumentLayout
      */
     public function setCX($value, $unit = self::UNIT_EMU)
     {
@@ -187,8 +185,8 @@ class DocumentLayout
      * Get Document Layout cy
      *
      * @param float $value
-     * @param float $unit
-     * @return integer
+     * @param string $unit
+     * @return DocumentLayout
      */
     public function setCY($value, $unit = self::UNIT_EMU)
     {
@@ -202,6 +200,7 @@ class DocumentLayout
      * @param float $value
      * @param string $fromUnit
      * @param string $toUnit
+     * @return float
      */
     protected function convertUnit($value, $fromUnit, $toUnit)
     {