+++ src/PhpPresentation/Shape/Chart/Axis.php
@@ -18,7 +18,6 @@
 namespace PhpOffice\PhpPresentation\Shape\Chart;
 
 use PhpOffice\PhpPresentation\ComparableInterface;
-use PhpOffice\PhpPresentation\Shape\Chart\Gridlines;
 use PhpOffice\PhpPresentation\Style\Font;
 
 /**
@@ -28,6 +27,11 @@ class Axis implements ComparableInterface
 {
     const AXIS_X = 'x';
     const AXIS_Y = 'y';
+
+    const TICK_MARK_NONE = 'none';
+    const TICK_MARK_CROSS = 'cross';
+    const TICK_MARK_INSIDE = 'in';
+    const TICK_MARK_OUTSIDE = 'out';
 
     /**
      * Title
@@ -59,6 +63,36 @@ class Axis implements ComparableInterface
      * @var Gridlines
      */
     protected $minorGridlines;
+
+    /**
+     * @var int
+     */
+    protected $minBounds;
+
+    /**
+     * @var int
+     */
+    protected $maxBounds;
+
+    /**
+     * @var string
+     */
+    protected $minorTickMark = self::TICK_MARK_NONE;
+
+    /**
+     * @var string
+     */
+    protected $majorTickMark = self::TICK_MARK_NONE;
+
+    /**
+     * @var float
+     */
+    protected $minorUnit;
+
+    /**
+     * @var float
+     */
+    protected $majorUnit;
 
     /**
      * Create a new \PhpOffice\PhpPresentation\Shape\Chart\Axis instance
@@ -141,6 +175,42 @@ class Axis implements ComparableInterface
     }
 
     /**
+     * @return int|null
+     */
+    public function getMinBounds()
+    {
+        return $this->minBounds;
+    }
+
+    /**
+     * @param int|null $minBounds
+     * @return Axis
+     */
+    public function setMinBounds($minBounds = null)
+    {
+        $this->minBounds = is_null($minBounds) ? null : (int)$minBounds;
+        return $this;
+    }
+
+    /**
+     * @return int|null
+     */
+    public function getMaxBounds()
+    {
+        return $this->maxBounds;
+    }
+
+    /**
+     * @param int|null $maxBounds
+     * @return Axis
+     */
+    public function setMaxBounds($maxBounds = null)
+    {
+        $this->maxBounds = is_null($maxBounds) ? null : (int)$maxBounds;
+        return $this;
+    }
+
+    /**
      * @return Gridlines
      */
     public function getMajorGridlines()
@@ -173,6 +243,78 @@ class Axis implements ComparableInterface
     public function setMinorGridlines(Gridlines $minorGridlines)
     {
         $this->minorGridlines = $minorGridlines;
+        return $this;
+    }
+
+    /**
+     * @return string
+     */
+    public function getMinorTickMark()
+    {
+        return $this->minorTickMark;
+    }
+
+    /**
+     * @param string $pTickMark
+     * @return Axis
+     */
+    public function setMinorTickMark($pTickMark = self::TICK_MARK_NONE)
+    {
+        $this->minorTickMark = $pTickMark;
+        return $this;
+    }
+
+    /**
+     * @return string
+     */
+    public function getMajorTickMark()
+    {
+        return $this->majorTickMark;
+    }
+
+    /**
+     * @param string $pTickMark
+     * @return Axis
+     */
+    public function setMajorTickMark($pTickMark = self::TICK_MARK_NONE)
+    {
+        $this->majorTickMark = $pTickMark;
+        return $this;
+    }
+
+    /**
+     * @return float
+     */
+    public function getMinorUnit()
+    {
+        return $this->minorUnit;
+    }
+
+    /**
+     * @param float $pUnit
+     * @return Axis
+     */
+    public function setMinorUnit($pUnit = null)
+    {
+        $this->minorUnit = $pUnit;
+        return $this;
+    }
+
+    /**
+     * @return float
+     */
+    public function getMajorUnit()
+    {
+        return $this->majorUnit;
+    }
+
+    /**
+     * @param float $pUnit
+     * @return Axis
+     */
+    public function setMajorUnit($pUnit = null)
+    {
+        $this->majorUnit = $pUnit;
         return $this;
     }