+++ src/PhpPresentation/Shape/Group.php
@@ -21,7 +21,6 @@ use PhpOffice\PhpPresentation\AbstractShape;
 use PhpOffice\PhpPresentation\GeometryCalculator;
 use PHPOffice\PhpPresentation\ShapeContainerInterface;
 use PhpOffice\PhpPresentation\Shape\Drawing;
-use PhpOffice\PhpPresentation\Shape\Line;
 use PhpOffice\PhpPresentation\Shape\RichText;
 use PhpOffice\PhpPresentation\Shape\Table;
 
@@ -63,7 +62,7 @@ class Group extends AbstractShape implements ShapeContainerInterface
     /**
     * Get collection of shapes
     *
-    * @return \ArrayObject
+     * @return \ArrayObject|AbstractShape[]
     */
     public function getShapeCollection()
     {