+++ src/PhpPresentation/Writer/PowerPoint2007/LayoutPack/AbstractLayoutPack.php
@@ -19,6 +19,7 @@ namespace PhpOffice\PhpPresentation\Writer\PowerPoint2007\LayoutPack;
 
 /**
  * \PhpOffice\PhpPresentation\Writer\PowerPoint2007\LayoutPack
+ * @deprecated 0.7
  */
 abstract class AbstractLayoutPack
 {
@@ -190,14 +191,13 @@ abstract class AbstractLayoutPack
      * Find specific slide layout id.
      *
      * @param string $name
-     * @param int $masterId
      * @return int
      * @throws \Exception
      */
-    public function findLayoutId($name = '', $masterId = 1)
+    public function findLayoutId($name = '')
     {
         foreach ($this->layouts as $layoutId => $layout) {
-            if ($layout['name'] == $name && $layout['masterid'] == $masterId) {
+            if ($layout['name'] == $name) {
                 return $layoutId;
             }
         }
@@ -209,14 +209,13 @@ abstract class AbstractLayoutPack
      * Find specific slide layout name.
      *
      * @param int $idLayout
-     * @param int $masterId
      * @return int
      * @throws \Exception
      */
-    public function findLayoutName($idLayout = '', $masterId = 1)
+    public function findLayoutName($idLayout = null)
     {
         foreach ($this->layouts as $layoutId => $layout) {
-            if ($layoutId == $idLayout && $layout['masterid'] == $masterId) {
+            if ($layoutId == $idLayout) {
                 return $layout['name'];
             }
         }