+++ tests/PhpPresentation/Tests/AbstractShapeTest.php
@@ -17,6 +17,7 @@
 
 namespace PhpOffice\PhpPresentation\Tests;
 
+use PhpOffice\PhpPresentation\Shape\Placeholder;
 use PhpOffice\PhpPresentation\Slide;
 use PhpOffice\PhpPresentation\Shape\Hyperlink;
 use PhpOffice\PhpPresentation\Shape\RichText;
@@ -150,5 +151,58 @@ class AbstractShapeTest extends \PHPUnit_Framework_TestCase
         $this->assertInstanceOf('PhpOffice\\PhpPresentation\\AbstractShape', $object->setWidthAndHeight($value, $value));
         $this->assertEquals($value, $object->getWidth());
         $this->assertEquals($value, $object->getHeight());
+    }
+
+    public function testPlaceholder()
+    {
+        $object = new RichText();
+        $this->assertFalse($object->isPlaceholder(), 'Standard Shape should not be a placeholder object');
+        $this->assertNull($object->getPlaceholder());
+        $this->assertInstanceOf(
+            'PhpOffice\\PhpPresentation\\AbstractShape',
+            $object->setPlaceHolder(new Placeholder(Placeholder::PH_TYPE_TITLE))
+        );
+        $this->assertTrue($object->isPlaceholder());
+        $this->assertInstanceOf('PhpOffice\\PhpPresentation\\Shape\\Placeholder', $object->getPlaceholder());
+        $this->assertEquals('title', $object->getPlaceholder()->getType());
+
+        $object = new RichText();
+        $this->assertFalse($object->isPlaceholder(), 'Standard Shape should not be a placeholder object');
+        $placeholder = new Placeholder(Placeholder::PH_TYPE_TITLE);
+        $placeholder->setType(Placeholder::PH_TYPE_SUBTITLE);
+        $this->assertInstanceOf('PhpOffice\\PhpPresentation\\AbstractShape', $object->setPlaceHolder($placeholder));
+        $this->assertTrue($object->isPlaceholder());
+        $this->assertInstanceOf('PhpOffice\\PhpPresentation\\Shape\\Placeholder', $object->getPlaceholder());
+        $this->assertEquals('subTitle', $object->getPlaceholder()->getType());
+    }
+
+    public function testContainer()
+    {
+        $object = new RichText();
+        $object2 = new RichText();
+        $object2->setWrap(RichText::WRAP_NONE);
+        $oSlide = new Slide();
+        $oSlide->addShape($object2);
+
+        $this->assertNull($object->getContainer());
+        $this->assertInstanceOf('PhpOffice\\PhpPresentation\\AbstractShape', $object->setContainer($oSlide));
+        $this->assertInstanceOf('PhpOffice\\PhpPresentation\\Slide', $object->getContainer());
+        $this->assertInstanceOf('PhpOffice\\PhpPresentation\\AbstractShape', $object->setContainer(null, true));
+        $this->assertNull($object->getContainer());
+    }
+
+    /**
+     * @expectedException \Exception
+     * @expectedExceptionMessage A \PhpOffice\PhpPresentation\ShapeContainerInterface has already been assigned. Shapes can only exist on one \PhpOffice\PhpPresentation\ShapeContainerInterface.
+     */
+    public function testContainerException()
+    {
+        $object = new RichText();
+        $oSlide = new Slide();
+
+        $this->assertNull($object->getContainer());
+        $this->assertInstanceOf('PhpOffice\\PhpPresentation\\AbstractShape', $object->setContainer($oSlide));
+        $this->assertInstanceOf('PhpOffice\\PhpPresentation\\Slide', $object->getContainer());
+        $this->assertInstanceOf('PhpOffice\\PhpPresentation\\AbstractShape', $object->setContainer(null));
     }
 }